requirements.txt
pandas>=1.3.0
scikit-learn>=1.0.0
numpy>=1.21.0
flask>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.25.0
Pillow>=8.0.0
cloudinary>=1.30.0
